FEAT/주문 생성에 SAGA 패턴 적용 (#282)
* refactor: 주문 수량(quantity) 관련 타입을 Long으로 통일

* feat: 주문 생성 SAGA 패턴 처리를 위한 엔티티 추가

* refactor: OrderRepository save 메서드 반환값 제거

* feat: 주문 생성 SAGA 처리용 에러 코드 추가

* feat: 주문 생성 SAGA 단계 구현 (재고/쿠폰/결제)

* feat: 비동기 처리 활성화를 위한 Async 설정 추가

* feat: 주문 생성 SAGA 관련 Repository 구현

* feat: 주문 생성 SAGA Step 구성 설정 추가

* feat: 주문 생성 SAGA 이벤트 발행자 추가

* feat: 주문 생성 SAGA Orchestrator 구현

* feat: 주문 생성 SAGA Orchestrator 실행 이벤트 리스너 추가

* feat: 주문 생성 SAGA 이벤트 처리 및 서비스 로직 추가

* refactor: SAGA 엔티티 및 enum을 infrastructure 레이어로 이동

* refactor: SAGA 엔티티 및 enum을 application 레이어로 이동

* refactor: SAGA 재시도 최대 횟수 조정

* chore: @Async 비활성화

* refactor: SagaOrchestrator run 메서드의 트랜잭션 제거

+++ order-service/src/main/java/com/palja/order_service/application/dto/external/ProductRes.java
@@ -16,5 +16,5 @@ public class ProductRes {
     private UUID productId;
     private String productName;
     private BigDecimal price;
-    private int stockQuantity;
+    private Long stockQuantity;
 }
