+++ order-service/src/main/java/com/palja/order_service/application/port/ProductClient.java
@@ -11,10 +11,10 @@ public interface ProductClient {
     ProductRes getProduct(UUID productId);
 
     // 상품 재고 차감
-    void deductProductStock(UUID productId, int quantity);
+    void deductProductStock(UUID productId, Long quantity);
 
     // 상품 재고 복구
-    void restoreProductStock(UUID productId, int quantity);
+    void restoreProductStock(UUID productId, Long quantity);
 
     List<UUID> getProductIdsByCompanyUserId(UUID companyUserId);
 }
