+++ order-service/src/main/java/com/palja/order_service/infrastructure/external/dto/response/ProductDTO.java
@@ -16,7 +16,7 @@ public class ProductDTO {
     private UUID companyUserId;
     private String productName;
     private BigDecimal price;
-    private int stockQuantity;
+    private Long stockQuantity;
 
     public  ProductRes toResponse() {
         return ProductRes.builder()
